move survived scratchpad logic into inference
