+++ self_hosting_machinery/inference/inference_hf.py
@@ -14,10 +14,9 @@ from transformers import StoppingCriteria
 from transformers import StoppingCriteriaList
 from transformers.generation.streamers import TextStreamer
 
-from refact_scratchpads import ScratchpadHuggingfaceBase
-from refact_scratchpads import ScratchpadHuggingfaceCompletion
+from self_hosting_machinery.inference.scratchpad_hf import ScratchpadHuggingfaceBase
+from self_hosting_machinery.inference.scratchpad_hf import ScratchpadHuggingfaceCompletion
 from self_hosting_machinery.inference import InferenceBase
-from self_hosting_machinery.inference import modload
 from refact_scratchpads_no_gpu.stream_results import UploadProxy
 
 from refact_utils.scripts import env
