+++ self_hosting_machinery/inference/scratchpad_hf.py
